+++ apps/send-func/src/__mocks__/notification.ts
@@ -182,11 +182,6 @@ export const aSendHeaders = sendHeadersSchema.parse({
   ...aLollipopHeaders,
 });
 
-export const aSendHeadersWithoutSrc = sendHeadersSchema.parse({
-  "x-pagopa-cx-taxid": aFiscalCode,
-  ...aLollipopHeaders,
-});
-
 export const aCIEValidationdata = CIEValidationDataSchema.parse({
   mrtdData: {
     dg1: "dg1",