+++ apps/services-func/CHANGELOG.md
@@ -1,5 +1,17 @@
 # services-func
 
+## 1.0.6
+
+### Patch Changes
+
+- 46b42db: Fix return of getBlobAsTextWithError on BlobNotFound 404
+
+## 1.0.5
+
+### Patch Changes
+
+- 173dc56: Add migration toolkit
+
 ## 1.0.4
 
 ### Patch Changes